add middleware login_auth

+++ laravel/routes/web.php
@@ -35,60 +35,58 @@ Route::get('/login', function () {
     return view('login.login');
 });
     
-Route::get('/form-food-banten', 'FoodBantenController@index')->name('admin.form-banten');
+Route::get('/form-food-banten', 'FoodBantenController@index')->name('admin.form-banten')->middleware('login_auth');
 
-Route::get('/form-food-jabar', 'FoodJabarController@index')->name('admin.form-jabar');
+Route::get('/form-food-jabar', 'FoodJabarController@index')->name('admin.form-jabar')->middleware('login_auth');
 
-Route::get('/form-food-jateng', 'FoodJatengController@index')->name('admin.form-jateng');
+Route::get('/form-food-jateng', 'FoodJatengController@index')->name('admin.form-jateng')->middleware('login_auth');
 
-Route::get('/form-food-jatim', 'FoodJatimController@index')->name('admin.form-jatim');
+Route::get('/form-food-jatim', 'FoodJatimController@index')->name('admin.form-jatim')->middleware('login_auth');
 
 
 Route::get('/login', 'AdminController@index')->name('login.index');
 Route::get('/logout', 'AdminController@logout')->name('login.logout');
 Route::post('/login', 'AdminController@process')->name('login.process');
 
-Route::get('/index', function () {
-    return view('admin.index');
-});
+Route::get('/index', 'IndexAdminController@index')->name('admin.index')->middleware('login_auth');
 
 
-Route::get('create-food-banten', 'FoodBantenController@create')->name('admin.create');
-Route::post('/postbanten', 'FoodBantenController@store')->name('admin.storeBanten');
-Route::get('/adminBanten/{id}/edit-food-banten', 'FoodBantenController@edit')->name('admin.editBanten');
-Route::patch('/adminBanten/{id}', 'FoodBantenController@update')->name('admin.updateBanten');
-Route::delete('/adminBanten/{id}', 'FoodBantenController@destroy')->name('admin.destroyBanten');
+Route::get('create-food-banten', 'FoodBantenController@create')->name('admin.create')->middleware('login_auth');
+Route::post('/postbanten', 'FoodBantenController@store')->name('admin.storeBanten')->name('admin.create')->middleware('login_auth');
+Route::get('/adminBanten/{id}/edit-food-banten', 'FoodBantenController@edit')->name('admin.editBanten')->middleware('login_auth');
+Route::patch('/adminBanten/{id}', 'FoodBantenController@update')->name('admin.updateBanten')->middleware('login_auth');
+Route::delete('/adminBanten/{id}', 'FoodBantenController@destroy')->name('admin.destroyBanten')->middleware('login_auth');
 
-Route::get('create-food-jabar', 'FoodJabarController@create')->name('admin.create');
-Route::post('/postjabar', 'FoodJabarController@store')->name('admin.storeJabar');
-Route::get('/adminJabar/{id}/edit-food-jabar', 'FoodJabarController@edit')->name('admin.editJabar');
-Route::patch('/adminJabar/{id}', 'FoodJabarController@update')->name('admin.updateJabar');
-Route::delete('/adminJabar/{id}', 'FoodJabarController@destroy')->name('admin.destroyJabar');
+Route::get('create-food-jabar', 'FoodJabarController@create')->name('admin.create')->middleware('login_auth');
+Route::post('/postjabar', 'FoodJabarController@store')->name('admin.storeJabar')->middleware('login_auth');
+Route::get('/adminJabar/{id}/edit-food-jabar', 'FoodJabarController@edit')->name('admin.editJabar')->middleware('login_auth');
+Route::patch('/adminJabar/{id}', 'FoodJabarController@update')->name('admin.updateJabar')->middleware('login_auth');
+Route::delete('/adminJabar/{id}', 'FoodJabarController@destroy')->name('admin.destroyJabar')->middleware('login_auth');
 
-Route::get('create-food-jateng', 'FoodJatengController@create')->name('admin.create');
-Route::post('/postjateng', 'FoodJatengController@store')->name('admin.storeJateng');
-Route::get('/adminJateng/{id}/edit-food-jateng', 'FoodJatengController@edit')->name('admin.editJateng');
-Route::patch('/adminJateng/{id}', 'FoodJatengController@update')->name('admin.updateJateng');
-Route::delete('/adminJateng/{id}', 'FoodJatengController@destroy')->name('admin.destroyJateng');
+Route::get('create-food-jateng', 'FoodJatengController@create')->middleware('login_auth');
+Route::post('/postjateng', 'FoodJatengController@store')->name('admin.storeJateng')->middleware('login_auth');
+Route::get('/adminJateng/{id}/edit-food-jateng', 'FoodJatengController@edit')->name('admin.editJateng')->middleware('login_auth');
+Route::patch('/adminJateng/{id}', 'FoodJatengController@update')->name('admin.updateJateng')->middleware('login_auth');
+Route::delete('/adminJateng/{id}', 'FoodJatengController@destroy')->name('admin.destroyJateng')->middleware('login_auth');
 
-Route::get('create-food-jatim', 'FoodJatimController@create')->name('admin.create');
-Route::post('/postjatim', 'FoodJatimController@store')->name('admin.storeJatim');
-Route::get('/adminJatim/{id}/edit-food-jatim', 'FoodJatimController@edit')->name('admin.editJatim');
-Route::patch('/adminJatim/{id}', 'FoodJatimController@update')->name('admin.updateJatim');
-Route::delete('/adminJatim/{id}', 'FoodJatimController@destroy')->name('admin.destroyJatim');
+Route::get('create-food-jatim', 'FoodJatimController@create')->middleware('login_auth');
+Route::post('/postjatim', 'FoodJatimController@store')->name('admin.storeJatim')->middleware('login_auth');
+Route::get('/adminJatim/{id}/edit-food-jatim', 'FoodJatimController@edit')->name('admin.editJatim')->middleware('login_auth');
+Route::patch('/adminJatim/{id}', 'FoodJatimController@update')->name('admin.updateJatim')->middleware('login_auth');
+Route::delete('/adminJatim/{id}', 'FoodJatimController@destroy')->name('admin.destroyJatim')->middleware('login_auth');
 
 
 
 Route::get('/adminJabar', 'FoodJabarController@index')
-->name('admin.indexJabar');
+->name('admin.indexJabar')->middleware('login_auth');
 
 Route::get('/adminBanten', 'FoodBantenController@index')
-->name('admin.indexBanten');
+->name('admin.indexBanten')->middleware('login_auth');
 
 Route::get('/adminJateng', 'FoodJatengController@index')
-->name('admin.indexJateng');
+->name('admin.indexJateng')->middleware('login_auth');
 
 Route::get('/adminJatim', 'FoodJatimController@index')
-->name('admin.indexJatim');
+->name('admin.indexJatim')->middleware('login_auth');
